+++ src/main/java/com/Soo_Shinsa/dto/ReviewRequestDto.java
@@ -1,4 +1,4 @@
-package com.Soo_Shinsa.dto.review;
+package com.Soo_Shinsa.dto;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
